slug naming for recipe url, rename cypress test files to avoid conflicting namespace, ignore cypress videos
