+++ cypress/e2e/gallery.spec.cy.ts
@@ -46,5 +46,3 @@ describe('Gallery page', () => {
       .should('be.visible')
   })
 })
-
-export {}
